Update README.md
HUNGAMA is a simple and fun command-line word guessing game written in Python, inspired by the classic Hangman. The game randomly selects a word from a predefined list and the player has to guess it one letter at a time, with a maximum of 6 incorrect guesses allowed. As the player inputs letters, the game reveals the correctly guessed ones in their positions while keeping the rest hidden. It also checks for repeated or invalid inputs and provides feedback after each guess. The game continues until the player successfully guesses all the letters or runs out of tries. It’s a beginner-friendly project that demonstrates basic Python concepts such as loops, conditionals, input validation, and list operations.